requirements.txt
google-search-results>=2.4.0
pandas>=1.3.0
tldextract>=3.4.0
typer[all]>=0.7.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.8.0
requests>=2.25.0
beautifulsoup4>=4.9.0
python-dotenv>=0.19.0
rich>=12.0.0
openpyxl>=3.0.0